change in student api and subject api

+++ src/services/api/StudentApi.tsx
@@ -12,37 +12,41 @@ export const StudentApi = createApi({
   baseQuery: fetchBaseQuery({
     baseUrl: "http://localhost:8000",
   }),
+
   tagTypes: ["readStudents", "readStudentsById"],
 
   endpoints: (builder) => ({
     createStudent: builder.mutation({
       query: (body) => {
+        const token = localStorage.getItem("token");
+        if (!token) {
+          throw new Error("No token available");
+        }
         return {
           url: "/students",
           method: "POST",
           body: body,
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
         };
       },
       invalidatesTags: ["readStudents"],
     }),
 
-    // UpdateStudent: builder.mutation({
-    //   query: (data) => {
-    //     return {
-    //       url: `/students/${data.id}`,
-    //       method: "PATCH",
-    //       body: data.body,
-    //     };
-    //   },
-    //   invalidatesTags: ["readStudents", "readStudentsById"],
-    // }),
-
     updateStudent: builder.mutation({
       query: (data) => {
+        const token = localStorage.getItem("token");
+        if (!token) {
+          throw new Error("No token available");
+        }
         return {
           url: `/students/${data.id}`,
           method: "PATCH",
           body: data.body,
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
         };
       },
       invalidatesTags: ["readStudents", "readStudentsById"],
@@ -51,12 +55,18 @@ export const StudentApi = createApi({
     readStudents: builder.query({
       query: (query: IQuery) => {
         const { page, limit, findQuery, sort } = query;
-
+        const token = localStorage.getItem("token");
+        if (!token) {
+          throw new Error("No token available");
+        }
         return {
           url: `/students?page=${page}&limit=${limit}&query=${encodeURIComponent(
             findQuery
           )}&sort=${sort}`,
           method: "GET",
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
         };
       },
       providesTags: ["readStudents"],
@@ -64,9 +74,16 @@ export const StudentApi = createApi({
 
     readStudentById: builder.query({
       query: (id) => {
+        const token = localStorage.getItem("token");
+        if (!token) {
+          throw new Error("No token available");
+        }
         return {
           url: `/students/${id}`,
           method: "GET",
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
         };
       },
       providesTags: ["readStudents"],
@@ -74,9 +91,16 @@ export const StudentApi = createApi({
 
     deleteStudent: builder.mutation({
       query: (id) => {
+        const token = localStorage.getItem("token");
+        if (!token) {
+          throw new Error("No token available");
+        }
         return {
           url: `/students/${id}`,
           method: "DELETE",
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
         };
       },
       invalidatesTags: ["readStudents"],
